Create LICENSE
I didn't know whether we should chose GPL or MIT.
Let's just assume people have good intentions, and not overly restrict usage by using a copyleft licence. 
The risk is someone hijacking the project (e.g. a well-funded company forking the project, then building a direct competitor on top of it), but  our choices shouldn't be driven by the worst possible outcome, they should be driven by the best possible outcome, which is that a lot of creative people use our project and take it in a successful direction we couldn't have seen ourselves